+++ yatm_radio_network/api.lua
@@ -1,7 +1,12 @@
+local path_join = assert(foundation.com.path_join)
+
 --- @namespace yatm_radio_network
 
 --- @const radio_network: RadioNetwork
-yatm_radio_network.radio_network = yatm_radio_network.RadioNetwork:new()
+yatm_radio_network.radio_network = yatm_radio_network.RadioNetwork:new{
+  filename = path_join(core.get_worldpath(), "radio_network"),
+}
+yatm.radio_network = yatm_radio_network.radio_network
 
 core.register_on_mods_loaded(yatm_radio_network.radio_network:method("init"))
 nokore_proxy.register_globalstep(